Create DB and insert values to DB of Neighborhood, Neighbor and Post
Form in /publish inserts information onto the DB, also the initial table creation script is in resources (commented)

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/persistence/PostDao.java
@@ -4,5 +4,5 @@ import ar.edu.itba.paw.models.Post;
 import ar.edu.itba.paw.models.User;
 
 public interface PostDao {
-    Post create(final String title, final String description, final int neighborId);
+    Post create(final String title, final String description, final long neighborId);
 }
