+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/NeighborService.java
@@ -10,5 +10,5 @@ public interface NeighborService {
 
     Optional<User> findById(long id);
 
-    Neighbor createNeighbor(final String email, final String name, final String surname, final int neighborhoodId);
+    Neighbor createNeighbor(final String email, final String name, final String surname, final long neighborhoodId);
 }
