+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/PostService.java
@@ -5,5 +5,5 @@ import ar.edu.itba.paw.models.Neighborhood;
 import ar.edu.itba.paw.models.Post;
 
 public interface PostService {
-    Post createPost(final String title, final String description, final int neighborId);
+    Post createPost(final String title, final String description, final long neighborId);
 }
